dynamic product paths
